// src/types/theme.ts
export type ThemeColor = {
  background: string;
  foreground: string;
  card: string;
  cardForeground: string;
  popover: string;
  popoverForeground: string;
  primary: string;
  primaryForeground: string;
  secondary: string;
  secondaryForeground: string;
  muted: string;
  mutedForeground: string;
  accent: string;
  accentForeground: string;
  destructive: string;
  destructiveForeground: string;
  border: string;
  input: string;
  ring: string;
};

export type Theme = {
  name: string;
  id: string;
  light: ThemeColor;
  dark: ThemeColor;
};

export type ThemeMode = 'light' | 'dark';